controller functions completed

+++ controllers/tourPackageController.js
@@ -2,6 +2,8 @@ const sequelize = require("../utils/database");
 const path = require("path");
 const fs = require("fs");
 const TourPackage = require("../models/tourPackageSchema");
+const TourTourPackageRelation = require("../models/tour-tourPackageRelation");
+const Tour = require("../models/tourSchema");
 const upload = require("../middleware/imageUploads"); // Adjust the path to your upload middleware
 
 //CREATE TOUR PACKAGE
@@ -40,49 +42,38 @@ exports.createTourPackage = async (req, res) => {
         });
       }
 
-      // Find existing tour IDs in the package
-      const existingTours = await TourPackage.findAll({
-        where: {
-          tourId: toursArray,
+      // Validate that tourPackageTitle already available
+      if (
+        await TourPackage.findOne({ where: { packageTitle: tourPackageTitle } })
+      ) {
+        await transaction.rollback();
+        return res.status(400).json({
+          success: false,
+          message: "Duplicate Tour Package Can't Created",
+        });
+      }
+
+      // Create the new TourPackage
+      const newPackage = await TourPackage.create(
+        {
           packageTitle: tourPackageTitle,
+          backgroundImage: tourPackageBgImageName,
         },
-        transaction,
-      });
-
-      const existingTourIds = existingTours.map((tour) => tour.tourId);
-
-      // Filter out existing tour IDs from the toursArray
-      const newToursArray = toursArray.filter(
-        (tourId) => !existingTourIds.includes(tourId)
+        { transaction }
       );
 
-      // If there are no new tours to add, return an appropriate message
-      if (newToursArray.length === 0) {
-        await transaction.rollback();
-        return res.status(400).json({
-          success: false,
-          message: "Some Tours already added to the package.",
-        });
+      // Create the relationships in the through table
+      for (const tourId of toursArray) {
+        await TourTourPackageRelation.create(
+          {
+            TourPackageId: newPackage.id,
+            TourId: tourId,
+          },
+          { transaction }
+        );
       }
-      // Map the tourPackageData for new tours
-      const tourPackageData = newToursArray.map((tourId) => ({
-        packageTitle: tourPackageTitle,
-        backgroundImage: tourPackageBgImageName,
-        tourId,
-      }));
 
-      try {
-        // Bulk create the new tour packages
-        await TourPackage.bulkCreate(tourPackageData, { transaction });
-        // Commit the transaction
-        await transaction.commit();
-      } catch (error) {
-        await transaction.rollback();
-        return res.status(400).json({
-          success: false,
-          message: "Tour not found",
-        });
-      }
+      await transaction.commit();
 
       res.status(200).json({
         success: true,
@@ -99,6 +90,62 @@ exports.createTourPackage = async (req, res) => {
   });
 };
 
+//GET ALL PACKAGES
+exports.getAllPackageTours = async (req, res) => {
+  try {
+    const allPackageTours = await TourPackage.findAll();
+    res.status(200).json({ success: true, allPackageTours });
+  } catch (error) {
+    console.error("Error retrieving all package tours:", error);
+    res.status(500).json({
+      success: false,
+      message: "An error occurred while retrieving package tours",
+    });
+  }
+};
+
+//GET ALL PACKAGES BY PACKAGE TOUR ID
+exports.getAllPackageToursAndTours = async (req, res) => {
+  const { tourPackageId, tourId } = req.query;
+
+  if (!tourPackageId || !tourId) {
+    return res.status(400).json({
+      success: false,
+      message: "Tour package ID or Tour ID is required",
+    });
+  }
+
+  try {
+    const packageWithTours = await TourTourPackageRelation.findAll({
+      where: { TourPackageId: tourPackageId, TourId: tourId },
+    });
+
+    if (!packageWithTours || packageWithTours.length === 0) {
+      return res.status(404).json({
+        success: false,
+        message: "Tour package not found",
+      });
+    }
+
+    const tour = await Tour.findByPk(packageWithTours[0].TourId);
+
+    if (!tour) {
+      return res.status(404).json({
+        success: false,
+        message: "Tour not found",
+      });
+    }
+
+    res.status(200).json({ success: true, packageWithTours, tour });
+  } catch (error) {
+    console.error("Error retrieving package tours and tours:", error);
+    res.status(500).json({
+      success: false,
+      message: "An error occurred while retrieving package tours and tours",
+    });
+  }
+};
+
 //DELETE TOUR PACKAGE
 exports.deleteTourPackage = async (req, res) => {
   const tourPackageId = req.params.id;
@@ -111,6 +158,7 @@ exports.deleteTourPackage = async (req, res) => {
 
   try {
     transaction = await sequelize.transaction();
+
     // Find the tour package by its ID
     const tourPackage = await TourPackage.findByPk(tourPackageId, {
       transaction,
@@ -120,14 +168,6 @@ exports.deleteTourPackage = async (req, res) => {
       await transaction.rollback();
       return res.status(404).send({ message: "Tour package not found" });
     }
-
-    // Find and delete all tour packages with the same packageTitle
-    const deleteCount = await TourPackage.destroy({
-      where: {
-        packageTitle: tourPackage.packageTitle,
-      },
-      transaction,
-    });
 
     // Delete associated images from server (if any)
     if (tourPackage.backgroundImage) {
@@ -139,25 +179,77 @@ exports.deleteTourPackage = async (req, res) => {
         tourPackage.backgroundImage
       );
 
-      fs.unlinkSync(imagePath); // Delete tour title image file from server
+      if (fs.existsSync(imagePath)) {
+        fs.unlinkSync(imagePath); // Delete tour package image file from server
+      }
     }
 
-    await transaction.commit();
-    console.log(
-      `Deleted ${deleteCount} tour packages with title: ${tourPackage.packageTitle}`
-    );
+    // Delete associated tour relationships
+    await TourTourPackageRelation.destroy({
+      where: { TourPackageId: tourPackageId },
+      transaction,
+    });
 
-    res
-      .status(200)
-      .send({ message: `Deleted ${deleteCount} tour packages successfully` });
+    // Delete the tour package itself
+    await TourPackage.destroy({
+      where: { id: tourPackageId },
+      transaction,
+    });
+
+    await transaction.commit();
+
+    res.status(200).send({ message: "Tour package deleted successfully" });
   } catch (error) {
-    await transaction.rollback();
-    console.error("Error deleting tour packages:", error);
-    res
-      .status(500)
-      .send({ message: "An error occurred while deleting tour packages" });
+    if (transaction) await transaction.rollback();
+    console.error("Error deleting tour package:", error);
+    res.status(500).send({
+      message: "An error occurred while deleting the tour package",
+    });
   }
 };
+
+//DELETE TOUR FROM TOUR PACKAGES
+exports.deleteTourFromPackage = async (req, res) => {
+  const { tourPackageId, tourId } = req.query;
+
+  if (!tourPackageId || !tourId) {
+    return res
+      .status(400)
+      .send({ message: "Invalid tour package ID or tour ID" });
+  }
+
+  let transaction;
+
+  try {
+    transaction = await sequelize.transaction();
+
+    // Check if the relationship exists
+    const relation = await TourTourPackageRelation.findOne({
+      where: { TourPackageId: tourPackageId, TourId: tourId },
+      transaction,
+    });
+
+    if (!relation) {
+      await transaction.rollback();
+      return res.status(404).send({ message: "Relation not found" });
+    }
+
+    // Delete the tour relationship
+    await relation.destroy({ transaction });
+
+    await transaction.commit();
+
+    res.status(200).send({ message: "Tour successfully removed from package" });
+  } catch (error) {
+    if (transaction) await transaction.rollback();
+    console.error("Error deleting tour from package:", error);
+    res.status(500).send({
+      message: "An error occurred while deleting the tour from the package",
+    });
+  }
+};
+
+// working in edit >>>>>>>>>>>>>>>>>
 
 //EDIT TOUR PACKAGE
 exports.editTourPackage = async (req, res) => {
